+++ modules/components/commands.js
@@ -1,5 +1,5 @@
 import { name, packageData, version } from "../constants.js";
-import { CommandInteraction, MessageEmbed } from "discord.js";
+import { CommandInteraction, EmbedBuilder } from "discord.js";
 import { discord, owners } from "../discord.js";
 import { log } from "../log.js";
 import { DateTime } from "luxon";
@@ -23,12 +23,14 @@ export const ping = async function(command) {
  * @param {CommandInteraction} command
  */
 export const about = async function(command) {
-    const embed = new MessageEmbed();
+    const embed = new EmbedBuilder();
     embed.setTitle(command.client.user.username);
     embed.setURL(`https://discord.com/api/oauth2/authorize?client_id=${command.client.user.id}&permissions=431644601408&scope=bot%20applications.commands`);
     embed.setDescription(`running [${name}](<${packageData.homepage}>) source code v${version}, for further info ${owners.length == 1 ? "contact" : "contact someone on this list:"} <@${owners.join(">, <@")}>`);
-    embed.addField("Commands", "`/ping`, `/about`, `/exit`, `/guilds`, `/save`, `/clear`", true);
-    embed.addField("Note", "the `/exit`, `/guilds`, `/save`, and `/clear` commands are restricted, and the latter two requires you have Manage Messages to appear as an option");
+    embed.addFields({
+        name: "Note",
+        value: "most commands are restricted and require you have Manage Messages to appear as an option",
+    });
     return await command.reply({
         embeds: [ embed ],
         ephemeral: true,
@@ -58,7 +60,7 @@ export const guilds = async function(command) {
     let list = command.client.guilds.cache.map((guild) => `${guild.name} (${guild.id})`).join("\n") || "no guilds?";
     log.debug({ guilds: list }, `${command.user.tag} (${command.user.id}) used /guilds`);
     if (list.length > 1900) {
-        list = list.substring(0, list.indexOf("\n", 1600));
+        list = list.substring(0, list.lastIndexOf("\n", 1900));
         list += `\nlist had to be truncated, see console or log file for the full list`;
     }
     await command.reply({
@@ -77,7 +79,7 @@ export const estimate = async function(command) {
     const ephemeral = command.options.getBoolean("ephemeral");
     const start = DateTime.fromMillis(0);
     const timeToFetch = start.plus({ seconds: Math.round(totalMessages / 100) });
-    let msg = null;
+    let msg = `unknown /estimate type`;
     if (type === "save") {
         const savingInterval = Interval.fromDateTimes(start, timeToFetch);
         msg = `saving would take at least ${humanizeDuration(savingInterval.length("milliseconds"))}`;