+++ modules/databases.js
@@ -7,7 +7,7 @@ import { directory } from "./constants.js";
  * lowdb json database used for analytics collected during recursively scanning
  * channels
  */
-export const analytics = new Low(new JSONFile(join(directory, "data", "data.json")));
+const analytics = new Low(new JSONFile(join(directory, "data", "data.json")));
 await analytics.read();
 if (!analytics.data) {
     analytics.data = {};
@@ -17,7 +17,7 @@ if (!analytics.data) {
 /**
  * lowdb json database used for saving emoji data when enabled
  */
-export const emojis = new Low(new JSONFile(join(directory, "data", "emojis.json")));
+const emojis = new Low(new JSONFile(join(directory, "data", "emojis.json")));
 await emojis.read();
 if (!emojis.data) {
     emojis.data = {};
@@ -25,3 +25,8 @@ if (!emojis.data) {
 }
 
 log.info("Started databases");
+
+export {
+    analytics,
+    emojis,
+};