+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-## 1.1.1
+## [1.2.0](https://github.com/06000208/jules/releases/tag/1.2.0)
 
-- put request delays back to 1000ms
+- made using the webhook optional
 - ensured the bot will crash on unhandled rejections, such as api errors like Invalid Webhook token
+- added support optional initial before and after bounds for iterating channels
+- updated dependencies
+- added quick and dirty support for json defined jobs
+- fixed message content intent
 
 ## [1.1.0](https://github.com/06000208/jules/releases/tag/1.1.0)